requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
python-telegram-bot==20.8
python-dotenv==1.0.1
webdriver-manager==4.0.1
requests==2.31.0
